+++ src/main/java/dev/ftb/mods/ftbic/screen/AntimatterConstructorMenu.java
@@ -5,7 +5,6 @@ import dev.ftb.mods.ftbic.block.entity.machine.AntimatterConstructorBlockEntity;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.DataSlot;
-import net.minecraft.world.inventory.Slot;
 
 public class AntimatterConstructorMenu extends ElectricBlockMenu {
 	public final DataSlot hasBoostSlot = DataSlot.standalone();
@@ -29,7 +28,7 @@ public class AntimatterConstructorMenu extends ElectricBlockMenu {
 		ElectricBlockEntityContainer container = new ElectricBlockEntityContainer(blockEntity);
 		int inputs = blockEntity.inputItems.length;
 		int outputs = blockEntity.outputItems.length;
-		if (inputs > 0) addSlot(new Slot(container, 0, 53, 36));
+		if (inputs > 0) addSlot(new FilteredInputSlot(container, 0, 53, 36));
 		if (outputs > 0) addSlot(new OutputSlot(container, inputs, 107, 36));
 		machineSlotCount = inputs + outputs;
 	}