+++ src/main/java/dev/ftb/mods/ftbic/screen/BatteryBoxMenu.java
@@ -3,7 +3,6 @@ package dev.ftb.mods.ftbic.screen;
 import dev.ftb.mods.ftbic.block.entity.ElectricBlockEntity;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.inventory.Slot;
 
 public class BatteryBoxMenu extends ElectricBlockMenu {
 	public BatteryBoxMenu(int id, Inventory playerInv, FriendlyByteBuf buf) {
@@ -23,7 +22,7 @@ public class BatteryBoxMenu extends ElectricBlockMenu {
 		ElectricBlockEntityContainer container = new ElectricBlockEntityContainer(blockEntity);
 		int inputs = blockEntity.inputItems.length;
 		int outputs = blockEntity.outputItems.length;
-		if (inputs > 0) addSlot(new Slot(container, 0, 53, 35));
+		if (inputs > 0) addSlot(new FilteredInputSlot(container, 0, 53, 35));
 		if (outputs > 0) addSlot(new OutputSlot(container, inputs, 109, 35));
 		machineSlotCount = inputs + outputs;
 	}